Created header component
