+++ src/components/Header.jsx
@@ -1,0 +1,14 @@
+function Header() {
+    return (
+      <header>
+        <img src="src/assets/react-core-concepts.png" alt="Stylized atom" />
+        <h1>React Essentials</h1>
+        <p>
+          Fundamental React concepts you will need for almost any app you are
+          going to build!
+        </p>
+      </header>
+    );
+  }
+
+  export default Header;
